fix: 解决arm64门禁编译问题
close: #I4ZQXT

+++ common/src/parser_util.cpp
@@ -57,7 +57,7 @@ int ParserUtil::ParserPdpProfileJson(std::vector<PdpProfile> &vec)
 
 void ParserUtil::ParserPdpProfileInfos(std::vector<PdpProfile> &vec, Json::Value &root)
 {
-    for (size_t i = 0; i < root.size(); i++) {
+    for (int i = 0; i < root.size(); i++) {
         Json::Value itemRoot = root[i];
         PdpProfile bean;
         bean.profileName = itemRoot[ITEM_OPERATOR_NAME].asString();
